requirements-dev.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
Django==2.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
